Move llm_config fixture to conftest
